asm: disallow labels in #bytes
The minimal size of a label depends on the output of the program, which in turn
depends on the sizes of #bytes instructions. The same problem exists for PUSH/JUMP,
and the compiler deals with it by messing with instruction sizes to discover the
optimal size. While this could perhaps be implemented for #bytes as well, it's not
really a necessary feature, so I'm opting to just disallow labels in #bytes.